+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## 0.1.17 - 2018-11-01
+### Changed
+* Fixes #41 - Use expound for sane error messages with spec
+* Fix #43 - Lots of useless output in 0.1.16
+* Fixes #38 - Use tools.cli v 0.4.1
+
+
 ## 0.1.16 - 2018-10-29
 ### Changed
 - Added Spec support for all options, and for the subcommand at once.